Restructure PhaseList and Phase test files to ease running separately
